Only delete Drafts

+++ src/api/exemptions/controllers/delete-exemption.js
@@ -3,6 +3,7 @@ import { StatusCodes } from 'http-status-codes'
 import { getExemption } from '../../../models/get-exemption.js'
 import { ObjectId } from 'mongodb'
 import { authorizeOwnership } from '../helpers/authorize-ownership.js'
+import { EXEMPTION_STATUS } from '../../../common/constants/exemption.js'
 
 export const deleteExemptionController = {
   options: {
@@ -15,18 +16,31 @@ export const deleteExemptionController = {
     try {
       const { params, db } = request
 
-      const result = await db
+      const exemption = await db
         .collection('exemptions')
-        .deleteOne({ _id: ObjectId.createFromHexString(params.id) })
+        .findOne({ _id: ObjectId.createFromHexString(params.id) })
 
-      if (result.deletedCount === 0) {
+      if (!exemption) {
         throw Boom.notFound('Exemption not found')
       }
+
+      if (exemption.status !== EXEMPTION_STATUS.DRAFT) {
+        throw Boom.badRequest(
+          `Cannot delete exemption as exemption must be the status '${EXEMPTION_STATUS.DRAFT}'.`
+        )
+      }
+
+      await db
+        .collection('exemptions')
+        .deleteOne({ _id: ObjectId.createFromHexString(params.id) })
 
       return h
         .response({ message: 'Exemption deleted successfully' })
         .code(StatusCodes.OK)
     } catch (error) {
+      if (error.isBoom) {
+        throw error
+      }
       throw Boom.internal(`Error deleting exemption: ${error.message}`)
     }
   }
